requirements.txt
streamlit
llama-index
llama-index-llms-groq
python-dotenv
torch 
torchvision 
torchaudio